% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{cond_mutual_inf}
\alias{cond_mutual_inf}
\title{Conditional mutual information estimation}
\usage{
cond_mutual_inf(X, Y, Z, k = 3L)
}
\arguments{
\item{X}{vector of size N.}

\item{Y}{input vector of length N or a matrix of size NxM.}

\item{Z}{conditional input vector of length N or a matrix of size NxM.}

\item{k}{number of nearest neighbors.}
}
\value{
The estimated conditional mutual information as a double-precision vector of length N.
}
\description{
Conditional mutual information estimation CMI(X;Y|Z) where X is a continuous vector.
The input Y and conditional input Z can be vectors or matrices. If Y and Z
are discrete then they must be numeric or integer valued.
}
\examples{
data(mutual_info_df)
set.seed(654321)
cond_mutual_inf(mutual_info_df$Zc_XcYc,
                       mutual_info_df$Xc, t(mutual_info_df$Yc))

M <- cbind(mutual_info_df$Xc, mutual_info_df$Yc)
ZM <- cbind(mutual_info_df$Yc, mutual_info_df$Wc)
cond_mutual_inf(mutual_info_df$Zc_XcYcWc, t(M), t(ZM))


}
\references{
Alkiviadis Tsimpiris, Ioannis Vlachos, Dimitris Kugiumtzis,
Nearest neighbor estimate of conditional mutual information in feature selection,
Expert Systems with Applications,
Volume 39, Issue 16, 2012, Pages 12697-12708
\doi{10.1016/j.eswa.2012.05.014}
}
