% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{cond_mutual_inf_ccc_2d}
\alias{cond_mutual_inf_ccc_2d}
\title{Conditional mutual information estimation for continuous case of a vector and matrix, given matrix.}
\usage{
cond_mutual_inf_ccc_2d(x, M, Z, k = 3L)
}
\arguments{
\item{x}{input vector.}

\item{M}{input matrix of the same length as x.}

\item{Z}{conditional input matrix of the same length as x.}

\item{k}{number of nearest neighbors.}
}
\value{
a double-precision vector
}
\description{
Compute conditional mutual information of vector\code{x}, matrix \code{M}
given matrix\code{Z}
where \code{x}, \code{M} and \code{Z} are all continuous
}
