% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{cond_mutual_inf_cdd_1d}
\alias{cond_mutual_inf_cdd_1d}
\title{Conditional mutual information estimation for continuous/discrete case of 3 vectors.}
\usage{
cond_mutual_inf_cdd_1d(x, y, z, k = 3L, seed = 0L)
}
\arguments{
\item{x}{input vector.}

\item{y}{input vector of the same length as x.}

\item{z}{conditional input vector of the same length as x.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed<=0 a random seed is used.}
}
\value{
a double-precision value - mutual information estimation for
vectors \code{x} and \code{y}, given \code{z}.
}
\description{
Compute conditional mutual information of \code{x},\code{y} given \code{z}
where \code{x} is continuous, \code{y} and \code{z} are discrete
}
\examples{
data(mutual_info_df)

cond_mutual_inf_cdd_1d(mutual_info_df$Xc,
                  mutual_info_df$Yd,
                  mutual_info_df$Zd_XdYd, k=3)
## 0.1754903

}
