% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{mutual_inf_cc_1d}
\alias{mutual_inf_cc_1d}
\title{Mutual information estimation for continuous case of 2 vectors.}
\usage{
mutual_inf_cc_1d(x, y, k = 3L, seed = 0L)
}
\arguments{
\item{x}{input vector.}

\item{y}{input vector of the same length as x.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed is <=0 a random seed is used.}
}
\value{
a double-precision value - mutual information estimation for
vectors \code{x} and \code{y}.
}
\description{
Compute mutual information of \code{x} and \code{y}
where \code{x} and \code{y} are both continuous
}
\examples{

x <- c(0.9065555 , 0.08110139, 0.13248763, 0.01142746, 0.07952208,
0.24536721, 0.86055117, 0.1314828 , 0.26211815, 0.69562545,
0.18115096, 0.69699724, 0.25942256, 0.2728219 , 0.45985588)

y <- c(1.75049046,  1.14456434,  0.99217277,  0.96046432,  1.08373406,
       1.17513354,  1.63132858,  1.02685152,  0.94713564, -0.31240944,
       0.88935441,  1.44502339,  1.02039948,  0.97471144,  1.25480345)


mutual_inf_cc_1d(x, y, k=3)

}
