% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{mutual_inf_cc_2d}
\alias{mutual_inf_cc_2d}
\title{Mutual information estimation for continuous case of an NxM matrix.}
\usage{
mutual_inf_cc_2d(x, M, k = 3L)
}
\arguments{
\item{x}{input vector.}

\item{M}{input matrix. The number of rows should be equal to the
length of vector x}

\item{k}{number of nearest neighbors.}
}
\value{
a vector of length equal to the number of columns of matrix M.
}
\description{
Compute conditional mutual information of \code{x} and matrix \code{M}
where \code{x} and \code{M} are both continuous
}
\examples{

set.seed(21)
x <- runif(10, min=0, max=1)

M = cbind(x + rnorm(10, 0, 0.1), x + rnorm(10, 0, 0.1),
                 x + rnorm(10, 0, 0.1))


mutual_inf_cc_2d(x, M, k=3)

}
