% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{mutual_inf_cd_2d}
\alias{mutual_inf_cd_2d}
\title{Mutual information estimation for continuous/discrete case.}
\usage{
mutual_inf_cd_2d(x, M, k = 3L, seed = 0L, use_cc = FALSE)
}
\arguments{
\item{x}{input (continuous) vector.}

\item{M}{input (discrete) matrix. It should have the same number of rows as \code{length(x)}.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed<=0 a random seed is used.}

\item{use_cc}{(logical) if TRUE the algorithm falls into continuous/continuous case}
}
\value{
vector of length m, where m is the number of columns in M
vectors \code{x} and \code{y}.
}
\description{
Compute mutual information of \code{x} and \code{y}
where \code{x} is continuous vector and \code{y} are discrete
}
