% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{cond_mutual_inf_ccc_1d}
\alias{cond_mutual_inf_ccc_1d}
\title{Conditional mutual information estimation for continuous case of 3 vectors.}
\usage{
cond_mutual_inf_ccc_1d(x, y, z, k = 3L)
}
\arguments{
\item{x}{input vector.}

\item{y}{input vector of the same length as x.}

\item{z}{conditional input vector of the same length as x.}

\item{k}{number of nearest neighbors.}
}
\value{
a double-precision value - mutual information estimation for
vectors \code{x} and \code{y}.
}
\description{
Compute conditional mutual information of \code{x},\code{y} given \code{z}
where \code{x}, \code{y} and \code{z} are all continuous
}
