% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnmi.R
\name{cond_mutual_inf_ccc_2d}
\alias{cond_mutual_inf_ccc_2d}
\title{Conditional mutual information estimation for continuous case of a vector and matrix, given matrix.}
\usage{
cond_mutual_inf_ccc_2d(x, M, Z, k = 3L, seed = 0L)
}
\arguments{
\item{x}{input vector.}

\item{M}{input matrix of the same length as x.}

\item{Z}{conditional input matrix of the same length as x.}

\item{k}{number of nearest neighbors.}

\item{seed}{integer random seed. If seed<=0 a random seed is used.}
}
\value{
a double-precision vector
}
\description{
Compute conditional mutual information of vector\code{x}, matrix \code{M}
given matrix\code{Z}
where \code{x}, \code{M} and \code{Z} are all continuous
}
\examples{
data(mutual_info_df)

M <- cbind(mutual_info_df$Xc, mutual_info_df$Yc)
cond_mutual_inf_ccc_2d(mutual_info_df$Zc_XcYcWc,
                  M,
                  mutual_info_df$Wc)
## 0.2029773 0.0000000

}
